pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tiantian-ai-guandan"
version = "0.3.0"
description = "Guandan card game AI agent"
requires-python = ">=3.9"
dependencies = [
    "opencv-python>=4.8",
    "mss>=9.0",
    "Pillow>=10.0",
    "scikit-image>=0.21",
    "numpy>=1.24",
    "easyocr>=1.7",
    "pynput>=1.7",
]

[project.scripts]
guandan = "guandan.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["guandan"]

[tool.coverage.report]
fail_under = 80
